# chore/cb latency optimization