--- a/src/bedrock_agentcore_starter_toolkit/operations/gateway/client.py
+++ b/src/bedrock_agentcore_starter_toolkit/operations/gateway/client.py
@@ -84,7 +84,7 @@ class GatewayClient:
             "protocolType": "MCP",
             "authorizerType": "CUSTOM_JWT",
             "authorizerConfiguration": authorizer_config,
-            "exceptionLevel": "DEBUG"
+            "exceptionLevel": "DEBUG",
         }
         if enable_semantic_search:
             create_request["protocolConfiguration"] = {"mcp": {"searchType": "SEMANTIC"}}