--- a/tests/services/test_codebuild.py
+++ b/tests/services/test_codebuild.py
@@ -1,7 +1,6 @@
 """Tests for Bedrock AgentCore CodeBuild service integration."""
 
 import json
-from datetime import timezone
 from unittest.mock import Mock, mock_open, patch
 
 import pytest
@@ -138,14 +137,10 @@ class TestCodeBuildService:
         mock_zip = Mock()
         mock_zipfile.return_value.__enter__.return_value = mock_zip
 
-        # Mock datetime for consistent timestamp
-        with patch("bedrock_agentcore_starter_toolkit.services.codebuild.datetime") as mock_dt:
-            mock_dt.now.return_value.strftime.return_value = "20250718-083000"
-            mock_dt.timezone = timezone
+        # Test with fixed source.zip naming (no timestamp needed)
+        result = codebuild_service.upload_source("test-agent")
 
-            result = codebuild_service.upload_source("test-agent")
-
-        expected_key = "test-agent/20250718-083000.zip"
+        expected_key = "test-agent/source.zip"
         expected_s3_url = f"s3://bedrock-agentcore-codebuild-sources-123456789012-us-west-2/{expected_key}"
 
         assert result == expected_s3_url
@@ -281,7 +276,7 @@ class TestCodeBuildService:
                 codebuild_service.wait_for_completion("test-build-id", timeout=1)
 
     def test_get_arm64_buildspec(self, codebuild_service):
-        """Test ARM64 buildspec generation - native build with parallel ECR auth."""
+        """Test ARM64 buildspec generation - native build with sequential ECR auth."""
         buildspec = codebuild_service._get_arm64_buildspec("test-ecr-uri")
 
         assert "version: 0.2" in buildspec
@@ -292,12 +287,14 @@ class TestCodeBuildService:
         assert "docker buildx build" not in buildspec
         assert "linux/arm64" not in buildspec
 
-        # Verify parallel operations
-        assert "docker build -t bedrock-agentcore-arm64 . &" in buildspec
+        # Verify sequential operations with validation
+        assert "Starting Docker build..." in buildspec
+        assert "docker build -t bedrock-agentcore-arm64 ." in buildspec
+        assert "Verifying Docker image was created..." in buildspec
+        assert "docker images bedrock-agentcore-arm64:latest" in buildspec
+        assert "Starting ECR authentication..." in buildspec
         assert "aws ecr get-login-password" in buildspec
-        assert "&" in buildspec  # Background execution
-        assert "wait" in buildspec  # Synchronization point
-        assert "Both build and auth completed successfully" in buildspec
+        assert "ECR authentication completed successfully" in buildspec
 
     def test_parse_dockerignore_existing_file(self, codebuild_service):
         """Test parsing existing .dockerignore file."""
@@ -403,12 +400,8 @@ node_modules
             mock_zip = Mock()
             mock_zipfile.return_value.__enter__.return_value = mock_zip
 
-            # Mock datetime
-            with patch("bedrock_agentcore_starter_toolkit.services.codebuild.datetime") as mock_dt:
-                mock_dt.now.return_value.strftime.return_value = "20250718-083000"
-                mock_dt.timezone = timezone
-
-                codebuild_service.upload_source("test-agent")
+            # Test with fixed source.zip naming
+            codebuild_service.upload_source("test-agent")
 
             # Verify correct files were included/excluded
             zip_calls = mock_zip.write.call_args_list
@@ -463,12 +456,8 @@ node_modules
             mock_zip = Mock()
             mock_zipfile.return_value.__enter__.return_value = mock_zip
 
-            # Mock datetime
-            with patch("bedrock_agentcore_starter_toolkit.services.codebuild.datetime") as mock_dt:
-                mock_dt.now.return_value.strftime.return_value = "20250718-083000"
-                mock_dt.timezone = timezone
-
-                codebuild_service.upload_source("test-agent")
+            # Test with fixed source.zip naming
+            codebuild_service.upload_source("test-agent")
 
             # Verify only non-ignored files were added to zip
             zip_calls = mock_zip.write.call_args_list