--- a/tests_integ/memory/test_create_memory.py
+++ b/tests_integ/memory/test_create_memory.py
@@ -1,9 +1,11 @@
+from random import randint
+
 from bedrock_agentcore import BedrockAgentCoreApp
 from bedrock_agentcore.memory import MemoryClient
-from random import randint
 
 app = BedrockAgentCoreApp()
 client = MemoryClient(region_name="us-west-2")
+
 
 @app.entrypoint
 def entrypoint(_payload):
@@ -12,9 +14,10 @@ def entrypoint(_payload):
 
     return {"memory_id": memory_id}
 
+
 def create_memory():
     name = "CustomerSupportAgentMemory" + str(randint(1, 10000))
-    description="Memory for customer support conversations"
+    description = "Memory for customer support conversations"
 
     memory = client.create_memory(
         name=name,
@@ -24,6 +27,7 @@ def create_memory():
     print(f"Memory ID: {memory.get('id')}")
     print(f"Memory: {memory}")
 
-    return memory.get('id')
+    return memory.get("id")
+
 
 app.run()