Improve first time code build latency from 103 to 50s. Internal doc: https://quip-amazon.com/9cgJAHoCaUqP/AgentCore-Launch-Performance-Investigation

```
(bedrock-agentcore-starter-toolkit) ➜  bedrock-agentcore-starter-toolkit git:(chore/cb_optimization) ✗ time uv run agentcore launch
🚀 Launching Bedrock AgentCore (codebuild mode - RECOMMENDED)...
   • Build ARM64 containers in the cloud with CodeBuild
   • No local Docker required (DEFAULT behavior)
   • Production-ready deployment

💡 Deployment options:
   • agentcore launch                → CodeBuild (current)
   • agentcore launch --local        → Local development
   • agentcore launch --local-build  → Local build + cloud deploy

Starting CodeBuild ARM64 deployment for agent 'test_siwabhi_8_30_125' to account 864899855746 (us-west-2)
Setting up AWS resources (ECR repository, execution roles)...
Getting or creating ECR repository for agent: test_siwabhi_8_30_125
Repository doesn't exist, creating new ECR repository: bedrock-agentcore-test_siwabhi_8_30_125
⠧ Launching Bedrock AgentCore...✅ ECR repository available: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125
Getting or creating execution role for agent: test_siwabhi_8_30_125
Using AWS region: us-west-2, account ID: 864899855746
Role name: AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
⠼ Launching Bedrock AgentCore...Role doesn't exist, creating new execution role: AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
Starting execution role creation process for agent: test_siwabhi_8_30_125
✓ Role creating: AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
Creating IAM role: AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
⠦ Launching Bedrock AgentCore...✓ Role created: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
⠏ Launching Bedrock AgentCore...✓ Execution policy attached: BedrockAgentCoreRuntimeExecutionPolicy-test_siwabhi_8_30_125
Role creation complete and ready for use with Bedrock AgentCore
✅ Execution role available: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
Preparing CodeBuild project and uploading source...
⠋ Launching Bedrock AgentCore...Getting or creating CodeBuild execution role for agent: test_siwabhi_8_30_125
Role name: AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
⠴ Launching Bedrock AgentCore...CodeBuild role doesn't exist, creating new role: AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
Creating IAM role: AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
⠙ Launching Bedrock AgentCore...✓ Role created: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
Attaching inline policy: CodeBuildExecutionPolicy to role: AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
⠸ Launching Bedrock AgentCore...✓ Policy attached: CodeBuildExecutionPolicy
Waiting for IAM role propagation...
⠏ Launching Bedrock AgentCore...CodeBuild execution role creation complete: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
⠦ Launching Bedrock AgentCore...Using .dockerignore with 44 patterns
⠏ Launching Bedrock AgentCore...Uploaded source to S3: test_siwabhi_8_30_125/source.zip
⠼ Launching Bedrock AgentCore...Created CodeBuild project: bedrock-agentcore-test_siwabhi_8_30_125-builder
Starting CodeBuild build (this may take several minutes)...
⠦ Launching Bedrock AgentCore...Starting CodeBuild monitoring...
⠇ Launching Bedrock AgentCore...🔄 QUEUED started (total: 0s)
⠙ Launching Bedrock AgentCore...✅ QUEUED completed in 1.1s
🔄 PROVISIONING started (total: 1s)
⠦ Launching Bedrock AgentCore...✅ PROVISIONING completed in 8.5s
🔄 DOWNLOAD_SOURCE started (total: 10s)
⠸ Launching Bedrock AgentCore...✅ DOWNLOAD_SOURCE completed in 2.1s
🔄 BUILD started (total: 12s)
⠋ Launching Bedrock AgentCore...✅ BUILD completed in 11.7s
🔄 POST_BUILD started (total: 23s)
⠏ Launching Bedrock AgentCore...✅ POST_BUILD completed in 6.4s
🔄 COMPLETED started (total: 30s)
⠹ Launching Bedrock AgentCore...✅ COMPLETED completed in 1.1s
🎉 CodeBuild completed successfully in 0m 30s
CodeBuild completed successfully
✅ CodeBuild project configuration saved
Deploying to Bedrock AgentCore...
⠸ Launching Bedrock AgentCore...✅ Agent created/updated: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ
Polling for endpoint to be ready...
⠙ Launching Bedrock AgentCore...Agent endpoint: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ/runtime-endpoint/DEFAULT
Deployment completed successfully - Agent: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ
✓ CodeBuild completed: bedrock-agentcore-test_siwabhi_8_30_125-builder:03a610c4-1644-4efc-8a92-a64a4baadf73
✓ ARM64 image pushed to ECR: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125:latest
╭─────────────────────────────────────────────────────────────────────────────────────────── CodeBuild Deployment Complete ───────────────────────────────────────────────────────────────────────────────────────────╮
│ CodeBuild ARM64 Deployment Successful!                                                                                                                                                                              │
│                                                                                                                                                                                                                     │
│ Agent Name: test_siwabhi_8_30_125                                                                                                                                                                                   │
│ CodeBuild ID: bedrock-agentcore-test_siwabhi_8_30_125-builder:03a610c4-1644-4efc-8a92-a64a4baadf73                                                                                                                  │
│ Agent ARN: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ                                                                                                                │
│ ECR URI: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125:latest                                                                                                                │
│                                                                                                                                                                                                                     │
│ ARM64 container deployed to Bedrock AgentCore.                                                                                                                                                                      │
│                                                                                                                                                                                                                     │
│ You can now check the status of your Bedrock AgentCore endpoint with:                                                                                                                                               │
│ agentcore status                                                                                                                                                                                                    │
│                                                                                                                                                                                                                     │
│ You can now invoke your Bedrock AgentCore endpoint with:                                                                                                                                                            │
│ agentcore invoke '{"prompt": "Hello"}'                                                                                                                                                                              │
│                                                                                                                                                                                                                     │
│ 📋 Agent logs available at:                                                                                                                                                                                         │
│    /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT/runtime-logs                                                                                                                            │
│    /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT/otel-rt-logs --log-stream-names otel-rt-logs                                                                                            │
│                                                                                                                                                                                                                     │
│ 💡 Tail logs with:                                                                                                                                                                                                  │
│    aws logs tail /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT/runtime-logs --follow                                                                                                     │
│    aws logs tail /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT/runtime-logs --since 1h                                                                                                   │
╰─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
uv run agentcore launch  2.12s user 0.51s system 5% cpu 50.580 total



----


Latency for second build improved from 83s to 35s.

```
(bedrock-agentcore-starter-toolkit) ➜  bedrock-agentcore-starter-toolkit git:(chore/cb_optimization) time uv run agentcore launch
🚀 Launching Bedrock AgentCore (codebuild mode - RECOMMENDED)...
   • Build ARM64 containers in the cloud with CodeBuild
   • No local Docker required (DEFAULT behavior)
   • Production-ready deployment

💡 Deployment options:
   • agentcore launch                → CodeBuild (current)
   • agentcore launch --local        → Local development
   • agentcore launch --local-build  → Local build + cloud deploy

Starting CodeBuild ARM64 deployment for agent 'test_siwabhi_8_30_125' to account 864899855746 (us-west-2)
Setting up AWS resources (ECR repository, execution roles)...
Using ECR repository from config: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125
Using execution role from config: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKRuntime-us-west-2-08cb4200dc
Preparing CodeBuild project and uploading source...
⠙ Launching Bedrock AgentCore...Using CodeBuild role from config: arn:aws:iam::864899855746:role/AmazonBedrockAgentCoreSDKCodeBuild-us-west-2-08cb4200dc
⠹ Launching Bedrock AgentCore...Using .dockerignore with 44 patterns
⠴ Launching Bedrock AgentCore...Uploaded source to S3: test_siwabhi_8_30_125/source.zip
Using CodeBuild project from config: bedrock-agentcore-test_siwabhi_8_30_125-builder
Starting CodeBuild build (this may take several minutes)...
⠋ Launching Bedrock AgentCore...Starting CodeBuild monitoring...
⠙ Launching Bedrock AgentCore...🔄 QUEUED started (total: 0s)
⠼ Launching Bedrock AgentCore...✅ QUEUED completed in 1.1s
🔄 PROVISIONING started (total: 1s)
⠧ Launching Bedrock AgentCore...✅ PROVISIONING completed in 7.4s
🔄 DOWNLOAD_SOURCE started (total: 9s)
⠋ Launching Bedrock AgentCore...✅ DOWNLOAD_SOURCE completed in 1.1s
🔄 PRE_BUILD started (total: 10s)
⠼ Launching Bedrock AgentCore...✅ PRE_BUILD completed in 1.1s
🔄 BUILD started (total: 11s)
⠦ Launching Bedrock AgentCore...✅ BUILD completed in 10.6s
🔄 POST_BUILD started (total: 21s)
⠹ Launching Bedrock AgentCore...✅ POST_BUILD completed in 5.3s
🔄 COMPLETED started (total: 27s)
⠦ Launching Bedrock AgentCore...✅ COMPLETED completed in 1.1s
🎉 CodeBuild completed successfully in 0m 27s
CodeBuild completed successfully
✅ CodeBuild project configuration saved
Deploying to Bedrock AgentCore...
⠦ Launching Bedrock AgentCore...✅ Agent created/updated: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ
Polling for endpoint to be ready...
⠏ Launching Bedrock AgentCore...Agent endpoint: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ/runtime-endpoint/DEFAULT
Deployment completed successfully - Agent: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ
✓ CodeBuild completed: bedrock-agentcore-test_siwabhi_8_30_125-builder:9c745287-6e5d-4aae-b753-2711be911bdb
✓ ARM64 image pushed to ECR: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125:latest
╭─────────────────────────────────────────────────────────────────────────────────────────── CodeBuild Deployment Complete ───────────────────────────────────────────────────────────────────────────────────────────╮
│ CodeBuild ARM64 Deployment Successful!                                                                                                                                                                              │
│                                                                                                                                                                                                                     │
│ Agent Name: test_siwabhi_8_30_125                                                                                                                                                                                   │
│ CodeBuild ID: bedrock-agentcore-test_siwabhi_8_30_125-builder:9c745287-6e5d-4aae-b753-2711be911bdb                                                                                                                  │
│ Agent ARN: arn:aws:bedrock-agentcore:us-west-2:864899855746:runtime/test_siwabhi_8_30_125-Y6JOlADguQ                                                                                                                │
│ ECR URI: 864899855746.dkr.ecr.us-west-2.amazonaws.com/bedrock-agentcore-test_siwabhi_8_30_125:latest                                                                                                                │
│                                                                                                                                                                                                                     │
│ ARM64 container deployed to Bedrock AgentCore.                                                                                                                                                                      │
│                                                                                                                                                                                                                     │
│ You can now check the status of your Bedrock AgentCore endpoint with:                                                                                                                                               │
│ agentcore status                                                                                                                                                                                                    │
│                                                                                                                                                                                                                     │
│ You can now invoke your Bedrock AgentCore endpoint with:                                                                                                                                                            │
│ agentcore invoke '{"prompt": "Hello"}'                                                                                                                                                                              │
│                                                                                                                                                                                                                     │
│ 📋 Agent logs available at:                                                                                                                                                                                         │
│    /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT                                                                                                                                         │
│    /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT --log-stream-names otel-rt-logs                                                                                                         │
│                                                                                                                                                                                                                     │
│ 💡 Tail logs with:                                                                                                                                                                                                  │
│    aws logs tail /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT --follow                                                                                                                  │
│    aws logs tail /aws/bedrock-agentcore/runtimes/test_siwabhi_8_30_125-Y6JOlADguQ-DEFAULT --since 1h                                                                                                                │
╰─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────╯
uv run agentcore launch  1.59s user 0.37s system 5% cpu 35.067 total
```


. 
```